pyproject.toml
[project]
name = "osint_system"
version = "0.1.0"
description = "LLM-powered multi-agent OSINT system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0",
    "langchain>=1.2.3",
    "langgraph>=1.0.5",
    "loguru>=0.7.3",
    "mcp>=1.25.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
]
